refactor: move API files to api/ folder
