+++ src/api/api.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 
-import { EmployersModule } from "src/employers/employers.module";
-import { JobsModule } from "src/jobs/jobs.module";
+import { EmployersModule } from "src/api/employers/employers.module";
+import { JobsModule } from "src/api/jobs/jobs.module";
 
 @Module({
   imports: [EmployersModule, JobsModule],
