+++ src/api/employers/infrastructure/persistence/entities/company.entity.spec.ts
@@ -1,4 +1,4 @@
-import { CompanyEntity } from "src/employers/infrastructure/persistence/entities/company.entity";
+import { CompanyEntity } from "src/api/employers/infrastructure/persistence/entities/company.entity";
 
 describe("CompanyEntity", () => {
   it("should be defined", () => {
