+++ src/api/jobs/application/transform/resources/offer.request.ts
