+++ src/api/jobs/application/transform/resources/offer.resource.spec.ts
