+++ src/api/jobs/domain/entities/offer.model.spec.ts
