+++ src/api/jobs/domain/entities/offer.model.ts
