+++ src/api/jobs/domain/repositories/offers.repository.ts
