+++ src/api/jobs/domain/services/offers.service.ts
