+++ src/api/jobs/infrastructure/persistence/entities/offer.entity.spec.ts
