+++ src/api/jobs/infrastructure/persistence/entities/offer.entity.ts
