+++ src/api/shared/communication/BaseResponse.ts
